Final Version

+++ backend/controllers/product.controller.js
@@ -1,83 +1,71 @@
-import Product from '../models/product.model.js';
+import mongoose from "mongoose";
+import Product from "../models/product.model.js";
 
-export const getProduct = async (req, res) => {
-  let products;
+export const getProducts = async (req, res) => {
   try {
-    products = await Product.find({});
+    const products = await Product.find({});
     res.status(200).json({ success: true, data: products });
   } catch (error) {
-    console.error('Error fetching products:', error);
-    return res.status(500).json({ message: 'Server Error' });
-  }
-};
-
-export const getProductById = async (req, res) => {
-  const { id } = req.params;
-  try {
-    const product = await Product.findById(id);
-    if (!product) {
-      return res.status(404).json({ message: 'Product not found' });
-    }
-    res.status(200).json({ success: true, data: product });
-  } catch (error) {
-    console.error('Error fetching product:', error);
-    return res.status(500).json({ message: 'Server Error' });
+    console.log("error in fetching products:", error.message);
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
 
 export const createProduct = async (req, res) => {
-  const product = req.body; //user input data
+  const product = req.body; // user will send this data
 
   if (!product.name || !product.price || !product.image) {
-    return res.status(400).json({ message: 'Please provide all fields' });
+    return res
+      .status(400)
+      .json({ success: false, message: "Please provide all fields" });
   }
 
   const newProduct = new Product(product);
+
   try {
     await newProduct.save();
-    res.status(201).json({
-      success: true,
-      message: 'Product created',
-      data: newProduct,
-    });
+    res.status(201).json({ success: true, data: newProduct });
   } catch (error) {
-    console.error('Error creating product:', error);
-    return res.status(500).json({ message: 'Server Error' });
+    console.error("Error in Create product:", error.message);
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
 
 export const updateProduct = async (req, res) => {
   const { id } = req.params;
-  const updates = req.body;
+
+  const product = req.body;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product Id" });
+  }
 
   try {
-    const product = await Product.findByIdAndUpdate(id, updates, { new: true });
-    if (!product) {
-      return res.status(404).json({ message: 'Product not found' });
-    }
-    res.status(200).json({
-      success: true,
-      message: 'Product updated',
-      data: product,
+    const updatedProduct = await Product.findByIdAndUpdate(id, product, {
+      new: true,
     });
+    res.status(200).json({ success: true, data: updatedProduct });
   } catch (error) {
-    console.error('Error updating product:', error);
-    return res.status(500).json({ message: 'Server Error' });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
-  console.log('Deleting product with id:', id);
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product Id" });
+  }
 
   try {
-    const product = await Product.findByIdAndDelete(id);
-    if (!product) {
-      return res.status(404).json({ message: 'Product not found' });
-    }
-    res.status(200).json({ success: true, message: 'Product deleted' });
+    await Product.findByIdAndDelete(id);
+    res.status(200).json({ success: true, message: "Product deleted" });
   } catch (error) {
-    console.error('Error deleting product:', error);
-    return res.status(500).json({ message: 'Server Error' });
+    console.log("error in deleting product:", error.message);
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
